# docs/openapi/openapi.yaml
openapi: 3.0.1
info:
    title: Microservice API Fleet
    description: Combined OpenAPI 3.0.1 specification for all microservices.
    version: 1.0.0
servers:
    - url: http://localhost:8000
...
paths:
...
    /{tenant_id}/infrastructure/{service_name}:
        get:
            tags:
                - tenant-service
            summary: GetServiceInfrastructure
            operationId: GetServiceInfrastructure
            x-category: public
            x-service: tenant-service
            x-source:
                - tenant-service/cmd/router.go:46
                - tenant-service/internal/handler/internal_tenant_handler.go:29
            parameters:
                - name: tenant_id
                  in: path
                  required: true
                  description: Path parameter tenant_id
                  schema:
                    type: string
                - name: service_name
                  in: path
                  required: true
                  description: Path parameter service_name
                  schema:
                    type: string
            responses:
                "200":
                    description: Successful response
...
    /api/payments/webhook/{provider}:
        post:
            tags:
                - payment-service
            summary: HandleWebhook
            operationId: HandleWebhook
            x-category: jwt
            x-service: payment-service
            x-source:
                - payment-service/cmd/router.go:23
                - payment-service/internal/handler/payment_handler.go:79
            security:
                - BearerAuth: []
            parameters:
                - name: provider
                  in: path
                  required: true
                  description: Path parameter provider
                  schema:
                    type: string
            responses:
                "200":
                    description: Successful response
...
    /credentials/setup:
        post:
            tags:
                - auth-service
            summary: SetupPassword
            operationId: SetupPassword
            x-category: public
            x-service: auth-service
            x-source:
                - auth-service/cmd/router.go:39
                - auth-service/internal/handler/auth_handler.go:73
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
            responses:
                "200":
                    description: Successful response
...
    /permissions/register:
        post:
            tags:
                - auth-service
            summary: RegisterPermissions
            operationId: RegisterPermissions
            x-category: public
            x-service: auth-service
            x-source:
                - auth-service/cmd/router.go:74
                - auth-service/internal/handler/internal_permission_handler.go:37
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
            responses:
                "200":
                    description: Successful response
    /refresh:
        post:
            tags:
                - auth-service
            summary: Refresh
            operationId: Refresh
            x-category: public
            x-service: auth-service
            x-source:
                - auth-service/cmd/router.go:42
                - auth-service/internal/handler/auth_handler.go:170
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
            responses:
                "200":
                    description: Successful response
...
    /select-tenant:
        post:
            tags:
                - auth-service
            summary: SelectTenant
            operationId: SelectTenant
            x-category: public
            x-service: auth-service
            x-source:
                - auth-service/cmd/router.go:41
                - auth-service/internal/handler/auth_handler.go:139
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
            responses:
                "200":
                    description: Successful response
    /setup-token:
        post:
            tags:
                - auth-service
            summary: CreateSetupToken
            operationId: CreateSetupToken
            x-category: public
            x-service: auth-service
            x-source:
                - auth-service/cmd/router.go:73
                - auth-service/internal/handler/internal_auth_handler.go:31
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
            responses:
                "200":
                    description: Successful response
...
components:
    securitySchemes:
        BearerAuth:
            bearerFormat: JWT
            scheme: bearer
            type: http
        InternalToken:
            in: header
            name: X-Internal-Service-Token
            type: apiKey